Stop tracking dist
It's not such a good idea after all. Now that the scripts are there to
ensure that releases have the latest dist we should not have the initial
problem that cause us to check this in. This is also motivated by the
fact that having dist checked in makes it hard to follow actual code
changes using gitk or git diff.