Add a toPEM method
The library checking the certificates needs a proper PEM certificate in
order to verify the signature of the message.
This commit is adding that method and renaming "normalizeCertificate" to
make it more consitent whith the new method.

I'm not 100% that this is how we go from X509 to PEM and vice versa but
this seems to be fine for now