Refactor service-provider.ts
Extract the functio that parses an xml string to the xml module so we
can reuse that for the Single Logout assertion.

Move the extract and response checks to a login response module - goal
would be to have a module by thing we support? I don't that sounds good
for now.