Allow passing a metadata file when creating a sp
This has some bigger breaking changes.

The main one is moving away from the constructor to have a create static
method. This enables us to force the initialization rather than have
potentially null attributes and to have some async operations done when
creating the instance.

The other one is more about making the options object more of a top
level one by putting optional parameters in a separate preferences
property. This should add some clarity and because I was already making
a breaking change it made sense to do that at the same time.

Feature wise we can now pass idp as a string which should be the
metadata xml and the service provider should infer what it needs from
there.