Return a value when the xml is not encrypted
It's easier to use return a value when the xpath does not return any
match as that enables us to not have to check if the assertion is
encrypted before calling this function. There might be a better way of
checking - maybe we can expect it from the metadata file or the options
but not having to worry about it is a nice enough feature for now.